rrr: convert base/logging to module (rrr.logging)
logging.cpp becomes the rrr.logging module interface unit; logging.hpp
deleted. The Log class + Log_* template wrappers + static methods all
live in one place now.

GMF forward-declares `time_now_str` (same global-module-attachment
trick used for `get_exec_path` in rrr.debugging) since base/misc isn't
a module yet. Removed the `Pthread_mutex_{lock,unlock}` wrappers from
threading.hpp and call `pthread_mutex_{lock,unlock}` directly — those
wrappers were the only logging dependency on threading and the verify
inside them is redundant with libc's guarantee for an
already-initialized mutex.

Includers updated (3):
  - src/rrr/rrr.hpp, src/rrr/base/all.hpp: drop the include, add the
    import at the bottom of the textual block.
  - src/rrr/rpc/utils.hpp: comment-only reference; no code change.

Measurement (rrr + rpcbench, -j32, clean):
  +debugging:  20.98 s, 102.0 MB PCM
  +logging:    21.64 s, 102.5 MB PCM, librrr.a 10.39 MB

+0.7 s, +0.5 MB BMI (logging's BMI is small — mostly the static class).
