+++ src/rrr/CMakeLists.txt
@@ -19,6 +19,7 @@ file(GLOB RRR_BENCH_SRC
 set(RRR_MODULE_SRC
     ${CMAKE_CURRENT_SOURCE_DIR}/base/basetypes.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/base/debugging.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/base/logging.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/base/strop.cpp
 )
 list(REMOVE_ITEM RRR_BENCH_SRC ${RRR_MODULE_SRC})
@@ -47,7 +48,7 @@ if(NOT TARGET rrr)
         ${CMAKE_CURRENT_SOURCE_DIR}/reactor/quorum_event.cc
         ${CMAKE_CURRENT_SOURCE_DIR}/reactor/epoll_wrapper.cc
         ${CMAKE_CURRENT_SOURCE_DIR}/reactor/reactor.cc
-        ${CMAKE_CURRENT_SOURCE_DIR}/base/logging.cpp
+        # base/logging.cpp omitted — module unit.
         ${CMAKE_CURRENT_SOURCE_DIR}/base/misc.cpp
         # base/basetypes.cpp omitted — module unit (see RRR_MODULE_SRC).
         # base/debugging.cpp omitted — module unit.
